Also redirect `openat` syscall
Old code handled `open(char *filename, ...)`.

Add support for `openat(int fd, char *filename, ...)`.

Filename argument is in a different register, so add `reg` parameter to
both the filename-reading and filename-modifying functions.

Modify `wait_for_open` to return `0` if it should exit, and the value
for `reg` if it's a handled syscall.

Modify BPF filter to trace `openat`.